+++ tests/test_pact_group.py
@@ -10,12 +10,16 @@ def test_group_wait_during(checkpoint, timed_group, num_seconds):
     assert checkpoint.called_times == num_seconds + 1
 
 
-def test_group_wait_then(checkpoint, checkpoint2, timed_group, timed_pact):
+def test_group_wait_then(checkpoint, checkpoint1, checkpoint2, checkpoint3, timed_group, timed_pact):
+    timed_group.lastly(checkpoint1)
     timed_group.then(checkpoint)
     timed_pact.then(checkpoint2)
+    timed_pact.lastly(checkpoint3)
     timed_group.wait()
     assert checkpoint.called
+    assert checkpoint1.called
     assert checkpoint2.called
+    assert checkpoint3.called
     assert timed_group.is_finished()
 
 
@@ -56,37 +60,47 @@ def test_group_and_pact_with_duration(pact_duration, num_seconds):
     assert group_duration == flux.current_timeline.time() - time_before
 
 
-def test_group_without_absorb_then(pred1, pred2, checkpoint1, checkpoint2):
-    p1 = Pact('a').until(pred1).then(checkpoint1)
-    p2 = Pact('b').until(pred2).then(checkpoint2)
+def test_group_without_absorb(pred1, pred2, checkpoint, checkpoint1, checkpoint2, checkpoint3):
+    p1 = Pact('a').until(pred1).lastly(checkpoint).then(checkpoint1)
+    p2 = Pact('b').until(pred2).then(checkpoint2).lastly(checkpoint3)
     group = p1 + p2
     assert not checkpoint1.called
     pred1.satisfy()
     assert not group.poll()
+    assert checkpoint.called
     assert checkpoint1.called
     assert not checkpoint2.called
+    assert not checkpoint3.called
     pred2.satisfy()
     assert group.poll() and group.is_finished()
     assert checkpoint2.called
+    assert checkpoint3.called
 
 
-def test_group_with_absorb_then(pred1, pred2, checkpoint1, checkpoint2):
-    p1 = Pact('a').until(pred1).then(checkpoint1)
-    p2 = Pact('b').until(pred2).then(checkpoint2)
+def test_group_with_absorb(pred1, pred2, checkpoint, checkpoint1, checkpoint2, checkpoint3):
+    p1 = Pact('a').until(pred1).lastly(checkpoint).then(checkpoint1)
+    p2 = Pact('b').until(pred2).then(checkpoint2).lastly(checkpoint3)
     group = PactGroup()
     group.add(p1, absorb=True)
     group.add(p2, absorb=True)
     assert not p1._then
+    assert not p1._lastly
     assert not p2._then
+    assert not p2._lastly
+    assert not checkpoint.called
     assert not checkpoint1.called
     pred1.satisfy()
     assert not group.poll() and not group.is_finished()
+    assert not checkpoint.called
     assert not checkpoint1.called
     assert not checkpoint2.called
+    assert not checkpoint3.called
     pred2.satisfy()
     assert group.poll() and group.is_finished()
+    assert checkpoint.called
     assert checkpoint1.called
     assert checkpoint2.called
+    assert checkpoint3.called
 
 
 def test_abosrbing_group_not_implemented():