+++ doc/changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* :feature:`-` Add lastly callbacks that are similar to 'then' but called afterwards
 * :release:`1.7.0 <27-06-2017>`
 * :feature:`18` Allow specifying overall deadlines for pacts and groups
 * :release:`1.6.0 <15-12-2016>`